% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdt-package.R
\docType{package}
\name{rsdt-package}
\alias{rsdt}
\alias{rsdt-package}
\title{rsdt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
