% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdt-package.R
\docType{package}
\name{rsdt-package}
\alias{rsdt}
\alias{rsdt-package}
\title{rsdt: What the Package Does (One Line, Title Case)}
\description{
Tidy interface to the UNHCR RSD/RST Statistical Reports.
}
\author{
\strong{Maintainer}: Hisham Galal \email{galal@unhcr.org}

Other contributors:
\itemize{
  \item UNHCR [copyright holder]
}

}
\keyword{internal}
